# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 'API Libri Usati Crema documentazione'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://api.libriusaticrema.it'
...
paths:
...
  /api/prenotazioni/rimuoviLibro:
    post:
      summary: 'Rimuovi libro prenotazione'
      operationId: rimuoviLibroPrenotazione
      description: "Rimuove l'assegnazione di un libro da una prenotazione."
      parameters: []
      responses: {  }
      tags:
        - Prenotazioni
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_libro:
                  type: integer
                  description: 'ID del libro da scollegare.'
                  example: null
              required:
                - id_libro
      security: []
  /api/prenotazioni/scambiaLibro:
    post:
      summary: 'Scambia libro prenotazione'
      operationId: scambiaLibroPrenotazione
      description: 'Sostituisce un libro già assegnato a una prenotazione con un altro libro disponibile.'
      parameters: []
      responses: {  }
      tags:
        - Prenotazioni
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vecchio_id:
                  type: integer
                  description: 'ID del libro da sostituire.'
                  example: null
                nuovo_id:
                  type: integer
                  description: 'ID del nuovo libro.'
                  example: null
              required:
                - vecchio_id
                - nuovo_id
      security: []
...
  '/api/prenotazioni/libriScambiabili/{isbn}':
    get:
      summary: 'Libri scambiabili'
      operationId: libriScambiabili
      description: 'Restituisce i libri disponibili per sostituire un libro prenotato.'
      parameters: []
      responses: {  }
      tags:
        - Prenotazioni
      security: []
    parameters:
      -
        in: path
        name: isbn
        description: "L'ISBN del libro di riferimento."
        example: null
        required: true
        schema:
          type: string
...